backend/User_Management/userProfiles/requirements.txt

python-dotenv
flask-cors
pyjwt
pymongo